requirements.txt
Pillow>=10.0.0
ImageHash>=4.3.1
streamlit>=1.28.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dateutil>=2.8.0
uuid